(PC-8357) Refacto: move the cancellation button in a dedicated component
